CoverageをInterface化
